+++ src/Relation/BelongsTo.php
@@ -11,7 +11,6 @@
 
 namespace Pulsar\Relation;
 
-use ICanBoogie\Inflector;
 use Pulsar\Model;
 use Pulsar\Query;
 
@@ -20,27 +19,6 @@ use Pulsar\Query;
  */
 final class BelongsTo extends AbstractRelation
 {
-    /**
-     * @param string $localKey     identifying key on local model
-     * @param string $foreignModel foreign model class
-     * @param string $foreignKey   identifying key on foreign model
-     */
-    public function __construct(Model $localModel, ?string $localKey, string $foreignModel, ?string $foreignKey)
-    {
-        if (!$foreignKey) {
-            $foreignKey = Model::DEFAULT_ID_NAME;
-        }
-
-        // the default local key would look like `user_id`
-        // for a model named User
-        if (!$localKey) {
-            $inflector = Inflector::get();
-            $localKey = strtolower($inflector->underscore($foreignModel::modelName())).'_id';
-        }
-
-        parent::__construct($localModel, $localKey, $foreignModel, $foreignKey);
-    }
-
     protected function initQuery(Query $query): Query
     {
         $id = $this->localModel->{$this->localKey};