+++ src/Relation/HasOne.php
@@ -11,7 +11,6 @@
 
 namespace Pulsar\Relation;
 
-use ICanBoogie\Inflector;
 use Pulsar\Exception\ModelException;
 use Pulsar\Model;
 use Pulsar\Query;
@@ -21,27 +20,6 @@ use Pulsar\Query;
  */
 final class HasOne extends AbstractRelation
 {
-    /**
-     * @param string $localKey     identifying key on local model
-     * @param string $foreignModel foreign model class
-     * @param string $foreignKey   identifying key on foreign model
-     */
-    public function __construct(Model $localModel, ?string $localKey, string $foreignModel, ?string $foreignKey)
-    {
-        // the default foreign key would look like
-        // `user_id` for a model named User
-        if (!$foreignKey) {
-            $inflector = Inflector::get();
-            $foreignKey = strtolower($inflector->underscore($localModel::modelName())).'_id';
-        }
-
-        if (!$localKey) {
-            $localKey = Model::DEFAULT_ID_NAME;
-        }
-
-        parent::__construct($localModel, $localKey, $foreignModel, $foreignKey);
-    }
-
     protected function initQuery(Query $query): Query
     {
         $id = $this->localModel->{$this->localKey};