+++ tests/Models/Category.php
@@ -3,15 +3,13 @@
 namespace Pulsar\Tests\Models;
 
 use Pulsar\Model;
-use Pulsar\Relation\Relationship;
 
 class Category extends Model
 {
     protected static $properties = [
         'name' => [],
         'posts' => [
-            'relation' => Post::class,
-            'relation_type' => Relationship::HAS_MANY,
+            'has_many' => Post::class,
         ],
     ];
 }