+++ tests/Models/Garage.php
@@ -3,19 +3,16 @@
 namespace Pulsar\Tests\Models;
 
 use Pulsar\Model;
-use Pulsar\Relation\Relationship;
-use Pulsar\Type;
 
 class Garage extends Model
 {
     protected static $properties = [
-        'person_id' => [
-            'type' => Type::INTEGER,
-            'relation_type' => Relationship::BELONGS_TO,
+        'person' => [
+            'belongs_to' => Person::class,
         ],
         'location' => [],
         'cars' => [
-            'relation_type' => Relationship::HAS_MANY,
+            'has_many' => Car::class,
         ],
     ];
 }