+++ CHANGELOG.md
@@ -10,6 +10,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Added `Model::dirty()` to check if a property has an unsaved value, and whether it has changed.
 - Added `Type::*`, `Property::*`, and `Relationship::*` constants to replace removed `Model` constants.
 - New method `Model::hasId()` will indicate when a model has a complete identifier.
+- Model definitions have new shortcuts to define relationships: `belongs_to`, `belongs_to_many`, `has_many`, and `has_one`.
 
 ### Changed
 - Make model internal properties private when possible.